server/requirements.txt
amqp==5.3.1
asgiref==3.9.1
babel==2.17.0
billiard==4.2.1
celery==5.5.3
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.2.4
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-timezone-field==7.1
djangorestframework==3.16.0
docopt==0.6.2
dotenv==0.9.9
imap-tools==1.11.0
kombu==5.5.4
num2words==0.5.14
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
qrcode==8.2
redis==6.2.0
reportlab==4.4.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
